+++ Plugin/Thunderstore/CHANGELOG.md
@@ -3,6 +3,9 @@
 - Extra fix for dropping stuff infront of janitor while he's grabbing an item.
 - Added second member of cleanup crew, the transporter.
 - Added cleaning drone gal.
+- Improved pathfinding completely in all CodeRebirth entities, this makes them able to fluidly use all fire exits, entrances, anything they need to reach you no matter where you are, should preform better than usual too.
+- Hazards despawn if they spawn on top of doors.
+- Changed the extended logging config to be a level of logging, keep it at 0 if you don't want unnecessary logs.
 
 ## v0.12.3